awstest_back/requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.12
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.9
pilkit==2.0
Pillow==7.2.0
django-storages==1.10.1